Avoid useless fadeOut anim for non-seen scrollbar
The scorllbar will not show out if there is no extended range in the view,
but in current design, the no-needed scrollbar will still start a fadeOut
animation which drawing nothing during the animation period.

So check if the scrollbar needed to be drawn before start the animation, it
would avoid useless fadeOut animation which cannot be seen by user.

Change-Id: I8304f11d8649f3b466b0c81e44a0f3691c37d20a
Signed-off-by: Shuo Gao <shuo.gao@intel.com>
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>