AudioManager: make swap volume buttons RTL aware
The layout of the volume panel is already RTL aware...the problem is
if it is in landscape and the buttons are swapped they go into opposite
direction then the volume panel.

So this swap should only happen on LTR languages and RTL should be left
like it is.

So both LTR and RTL have it logical.

Change-Id: Ie5fbe3085b40ad35814f53ed19736d7800dad78e
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>