Frameworks: add ability to send a custom intent after keyguard is finished
We have and will have some features which need to start intents after the user
has sucessfully entered the credentials. Add for the system public available method
to start an avtivity after the keyguard.

Add as well support on slim Action for it so that our features actually can use this.

Signed-off-by: Chet Kener <Cl3Kener@gmail.com>

Conflicts:
	core/java/com/android/internal/statusbar/IStatusBar.aidl
	core/java/com/android/internal/statusbar/IStatusBarService.aidl
	packages/SystemUI/src/com/android/systemui/statusbar/CommandQueue.java
	packages/SystemUI/src/com/android/systemui/statusbar/tv/TvStatusBar.java
	services/core/java/com/android/server/statusbar/StatusBarManagerService.java

Change-Id: If9b6cbb39f1dd0b3fdad29f3f7b331133d32e89e
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>