Don't enable screen until keyguard is ready
The home screen was shown before the keyguard was ready. Allowing the
user to navigate home without unlocking the phone.

WindowManagerService calls checkWaitingForWindowsLocked to determine
when it is safe to enable the screen. But checkWaitingForWindowsLocked
only check if the Keyguard is ready if there is a Window of type
TYPE_STATUS_BAR.

If there is no statusbar window, the WindowManagerService assumed that
the keyguard was ready.

This change will assume that the keyguard is NOT ready until there is a
statusbar window and mPolicy.isKeyguardDrawnLw() return true.

Downside to this patch is that it will not prevent home screen from
being shown before keyguard if we have displayed any boot messages
during boot. At the moment we only show boot messages during first boot
and after upgrades.

Change-Id: Ia3da149568f966c9d753f658d939946dfa2f3b46
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>
Signed-off-by: Chet <Cl3Kener@gmail.com>