Allow no applications in navring, like in cm11
For those of us that do not have Google installed, or otherwise not
wanting the navring, this patch enables "None" to be chosen for the
center of the navring. This was disallowed in lollipop for some
reason.

This does require changing a previous decision to always show the middle
target. I'm not sure if that's done just as a design decision, a Google
CTS requirement, or just simplicity. The previous commit[1] and gerrit
changeset[2] does not reveal this point as far as I can tell.

[1]:
https://github.com/CyanogenMod/android_frameworks_base/commit/b57efee4a97f956c3cf1b1e55a69cff01635d601
[2]: http://review.cyanogenmod.org/#/c/87163/

Change-Id: Ib700d6f5febaaa144af2b4490a81539830c67284
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>