libandroidfw: Fatal exception of dlfree often causes the system crashed
When the function 'getBuffer' are called at the same time by the threads,
mZipInflater will be 'delete' twice if mZipInflater isn't NULL,
that is 'mZipInflater = NULL' has not be reached,
at this time it will cause system server crashed.

Add AutoMutex at the beginning of the function to avoid getBuffer to
be called at the same time.

CRs-Fixed: 812950
Change-Id: I283b43a01b97ae6a149749ff2186efda9df4dc42
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>