doze: Make delay and duration of intent triggered dozes configurable
Devices that have IR sensors can trigger dozes due directly to user
action (gestures) and the user expects a more immediate response
to this actions because  the user may be looking directly at
the screen and expecting immediate feedback.

It makes sense to differentiate this from the pickup sensor values
and notification sensor values because of the different user focus
(directly acting on the phone).

Change-Id: I085701bb7e5c0e1feb632ce945d615c7ae4e24a5
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>