Reducing computation time while doing alarm re-batching
Each invocation of setImplLocked() in rebatchAllAlarmsLocked() still try to search if
any duplicated alarm existing in the list and remove it. It's not necessary for
rebatching because each alarm in rebatching procedure is mutually unique. Skipping this
to reduce computation time of rebatching will improve performance while AlarmManager.cancel()
is called, or application will suffer in ANR.

Change-Id: Idaba4b9b22c0553a20ed80932f058d8b363687d7
Signed-off-by: Charles Tsai <charles_tsai@htc.com>
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>