textclock: Update the time when performing a doze pulse
Prior to this commit if your phone is deep sleeping then the
TIME_TICK intents are not necessarily being delivered to the
clock until you exit dreaming.  This provides a poor UI
when the clock is shown for a doze pulse (ambient display).

For example, I often see the time change by 2-3 minutes (and
up to 5 of more) when I simply touch the screen when it is
doing a doze pulse.  That change is caused by the ticks
finally getting registered and but doesn't necessarily happen
while still dreaming.

This defeats one useful purpose of the ambient display which is
to use it to see the current time.

Change-Id: Ib6eb656b0aea21b09f2778c8138f39c4ab11337f
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>