base: Nat464: unregister network observer when we stop!
Otherwise we will continue to receive updates.  Each time the
network is brought up and down, that will leave one more observer.
Each observer will be telling the ConnectivityService to update
the link properties for the link which caues it to reevaluate all
the network connections on all the links.

If a network link is brought up and down frequently this can then
cause a massive amount of CPU and kills deep sleep.

Change-Id: If5dc89c7b6418d4b6c44ef149e0dd6dc5e25b2ee
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>